Add search functionality to WorkflowDefinitionList (#166)
* Add search functionality to WorkflowDefinitionList

The _searchString private variable was replaced by a SearchTerm property. The OnSearch method was updated to reflect this change and renamed to OnSearchTermChanged. A search text field was also added to the WorkflowDefinitionList view. Some import statements have been adjusted and several other minor changes implement in UI.

* Update Elsa.Api.Client package version

The version of the Elsa.Api.Client has been updated from 3.1.0-preview.1133 to 3.1.0-4618.1149.

* Update Elsa.Api.Client package version

The Elsa.Api.Client package version has been updated from 3.1.0-4618.1150 to 3.1.0-preview.1152. This change is made to use the latest preview version of the package.