Lazy Loading of Subgraphs for Workflow Instance Viewer (#231)
* Remove 'includeCompositeRoot' parameter in workflow services

This commit removes the 'includeCompositeRoot' parameter from calls to FindByDefinitionIdAsync, FindByIdAsync, and FindManyByIdAsync methods in RemoteWorkflowDefinitionService and IWorkflowDefinitionService. It also changes the Elsa.Api.Client reference from package to project in Elsa.Studio.Core project file.

* Update tool and package versions to 3.3.0

The tool version has been updated in ToolVersion.cs. This change is reflected in the base_version environment variable in the GitHub actions configuration file. Additionally, the version in the GitHub actions file is no longer hardcoded but now references this base_version variable, ensuring consistency and easier future updates.

* Remove flag in workflow definition lookup

* Update WorkflowInstanceViewer and related components for subgraph support

Added support for subgraph in WorkflowInstanceViewer and related components. Made necessary changes in various classes including FallbackDesignerProvider, WorkflowInstanceDesigner, and others to incorporate the support for subgraph. The modifications were also reflected in the IDiagramDesignerProvider interface and IDiagramDesignerService to accept a WorkflowSubgraph parameter. Adjusted PathChanged event handling accordingly.

* Add IgnoreChildActivitiesConverter and improve activity JSON conversion

A new JSON converter, IgnoreChildActivitiesConverter, has been introduced to handle specific serialization scenarios. On top of this, improvements have been made to the existing ActivityJsonConverter. This includes more robust handling of synthetic properties and the ability to ignore child activities during serialization. In DiagramDesignerWrapper, an addition has been made to facilitate the handling of embedded activities.

* Remove WorkflowSubgraph from various components and services

This commit removes the WorkflowSubgraph from multiple components and services across the project. It includes adjustments made to various files to replace the usage of WorkflowSubgraph. This has resulted in fewer parameters in the activity functions and cleaner code. It's a key revision to streamline the codebase and simplify the future development process.

* Refactor and enhance AddNewActivityAsync method

Refactored the AddNewActivityAsync method in FlowchartDesignerWrapper.razor.cs to improve readability and also enhanced the method by setting a new 'nodeId' for the added activity, making the activity aware of its parent and its own identity. This addition helps in better handling and identification of activity in the flowchart.

* Update path to Elsa.Api.Client project

The path to 'Elsa.Api.Client' in the solution file has been updated. It now points to the 'perf-5350' folder instead of 'main', to reflect its new location in the codebase.

* Update project references and port configurations

The commit includes changes in various files that primarily update the project reference of Elsa.Api.Client.csproj to a different branch and modify ASPNETCORE_URLS and Backend server URLs. Additionally, better variable naming for enhanced code readability and logic adjustments are implemented in the DiagramDesignerWrapper.razor.cs file.

* Update Backend URL in appsettings.json

In the Elsa Studio Host, the Backend URL in the appsettings.json file has been changed. The port number was updated from 5001 to 8001 in the URL, adjusting the endpoint that the application communicates with.

* Update Elsa.Api.Client package to version 3.2.0-rc2.1823

The Elsa.Api.Client package has been updated to version 3.2.0-rc2.1823, which is referenced directly instead of the earlier project reference. This change applies to the Elsa.Studio.Core project and the package version property.

* Update server ports, tool version and clean up code

The commit includes changes to the server ports specified in different configuration files. The version of the tool has been downgraded from 3.3.0 to 3.2.0 across multiple files. Unnecessary code, mainly concerning an unused tab index and panel change listener, has been removed from WorkflowInstanceWorkspace.

Unnecessary mentions of the `Elsa.Api.Client.Resources.WorkflowDefinitions.Models` namespace are also removed from various files. Other unnecessary redundancies, like the OnActivePanelIndexChanged function, are also omitted.

* Refactor StitchNodesRecursive method in DiagramDesignerWrapper

The method StitchNodesRecursive in DiagramDesignerWrapper has been refactored for better readability and clarity. A new variable `isCollection` was introduced to simplify the conditional assignment. Also, unnecessary curly brackets surrounding the loop body were removed to streamline the code.

* Add explanation for StitchNodesRecursive function

The new code provides an illustrative example of how the StitchNodesRecursive function operates in the file DiagramDesignerWrapper.razor.cs. It makes clear how the process of converting a graph model of nodes to a structured model where child activities are properties of parent properties, works.

* Refine comment in DiagramDesignerWrapper

This commit makes minor modification to a comment in the DiagramDesignerWrapper file in Elsa Studio Workflows. The update clarifies that child activities are properties of their specific parent activity, rather than generic parent properties.

* Refactor code formatting and improve error handling

Adjusted code clustering and exception handling in workflow modules. Main changes include adjusting code layout for better readability and enhancing error management in activity subgraph selection to prevent silent failures. This will improve code maintainability and robustness.

* Refactor StitchNodesRecursive method call

The StitchNodesRecursive method call and embeddedActivity assignment in the DiagramDesignerWrapper class are updated. They now work with selectedPortGraph instead of selectedActivityGraph. This change ensures the accurate assignment of the port to the specific activity graph.

* Refactor exception handling in DiagramDesignerWrapper

Exceptions for missing activity and port graphs in the method DiagramDesignerWrapper are refactored to clean up the code. Now, these exceptions are thrown directly within the assignment lines for both selectedActivityGraph and selectedPortGraph, reducing the number of lines and improving readability.

* Update Elsa.Studio-Core settings

The .DotSettings file of Elsa.Studio-Core solution has been updated. It now includes additional ReSharper settings related to code style and formatting. This should help maintain consistent coding styles and improve the readability of the code.

* Remove StitchNodesRecursive method

The StitchNodesRecursive method that was used to convert a graph model of nodes to a structured model has been removed. As part of the code restructuring, it was no longer needed. We now directly access the required items from the 'selectedActivityGraph' object. This simplifies the codebase and reduces complexity.

* Update Elsa.Api.Client package version

The version of the Elsa.Api.Client package has been updated from 3.2.0-rc2.1823 to 3.2.0-rc2.1834 to include the most recent changes and improvements. This update ensures that the application is making use of the latest version of the package.

* Update project references in solution files

This commits updates the project references across various solution files. It specifically replaces references from the "perf-5350" branch to the "main" branch in the Elsa.Studio.Core project.

* Remove duplicate IsReadOnly property definition

The IsReadOnly property was defined twice in the WorkflowDefinitionEditor component, causing redundancy. This commit removes the duplicate definition to clean up the code and avoid any potential conflicts or confusion.