Optimize rendering of ActivityWrapper.
This change prevents the contents of the MudBadge component to not be re-rendered, which ought to improve re-rendering of the custom element as a whole.

However, it will not fix the re-rendering issue that is reported, given the fact that the entire custom element instance will be recreated by X6, when it receives the `stats` value.