Refactor: Remove unnecessary Blazor service context invocations (#307)
* Refactor: Remove unnecessary Blazor service context invocations

This commit eliminates redundant usage of InvokeWithBlazorServiceContext, streamlining the service calls directly. It also refines several constructors and DI usages to be more concise, enhancing overall code readability and maintainability.

* Update base version in CI workflow to 3.2.2

This change updates the base_version in the packages.yml file from 3.2.1 to 3.2.2. This ensures that the CI pipeline uses the latest version for package releases.