Fix Monaco Editor Suggestion Overlay Cutoff (#470)
* Update Monaco editor config

Added `FixedOverflowWidgets` option to Monaco editor configuration for improved behavior.

* Update elsa-studio-wasm to version 3.4.0-rc1

Upgraded dependency "@elsa-workflows/elsa-studio-wasm" from ^3.0.0 to ^3.4.0-rc1 in the React wrapper package. This ensures compatibility with the latest features and fixes provided in the 3.4.0-rc1 release.

* Update @elsa-workflows/elsa-studio-wasm to version 3.4.0-rc1

Upgraded the dependency from version 3.0.0 to 3.4.0-rc1 in package-lock.json. This includes updated integrity hashes and aligns with the latest release candidate.