New Content Formatting Viewer (#549)
* Add content formatting system and UI integration

This commit introduces a new content formatting system with the creation of `IContentFormatProvider` and `IContentFormatter` interfaces. Implementations for generic and JSON formatting are provided in `DefaultContentFormatter` and `JsonContentFormatter`. The `DefaultContentFormatProvider` manages registered formatters.

Additionally, the `ServiceCollectionExtensions` class is updated to support dependency injection for the new formatters. UI components, including `ContentFormatter.razor` and `DataPanel.razor`, are modified to allow users to select and view formatted content. The project file is also updated for proper package references.

* Update Nuget Packages and remove unnecessary using's. Set MudTextField to ReadOnly.

* Update MudBlazor package version to 8.8.0

Updated the version of the `MudBlazor` package from `8.7.0` to `8.8.0` in the `Directory.Packages.props` file.

* Add Syntax property to IContentFormatter and Improve UX.

- Updated `IContentFormatter` to include a new `Syntax` property.
- Enhanced `ContentFormatter.razor` with a new toolbar and improved layout.
- Add Monaco Editor support for better prettified rendering.

* Enhance localization and UI in ContentFormatter

- Added localization support for tooltips and tab titles in ContentFormatter.razor.
- Improved read-only functionality in the editor with dynamic state management.
- Updated table rendering logic for better header and row display.
- Injected ILocalizer dependency in ContentFormatter.razor.cs.
- Enhanced copy action feedback and simplified success messages.
- Introduced CSS for multiline text fading effects in DataPanel.razor.
- Added truncationLength variable for better text display management.

* Remove fade-multiline class and update text rendering

This commit removes the commented-out CSS class `.fade-multiline` that controlled display properties for text elements. It also updates the rendering logic for items with text exceeding a specified truncation length, replacing the previous `<div class="fade-multiline">` with a simpler `<div>`, and adds a tooltip and link for viewing the full text.